pyproject.toml
[tool.poetry]
authors = ["Andrei Aksionov"]
description = "Semantic search powered by Weaviate vector search engine"
license = "MIT"
name = "weaviate-demo"
packages = [
  {include = "src"},
]
version = "0.1.0"

[tool.poetry.dependencies]
newspaper3k = "~0.2.8"
nltk = "~3.6.7"
pandas = "~1.4.0"
python = "~3.9.9"
sentence-transformers = "~2.2.0"
torch = "~1.10.2"
tqdm = "~4.62.3"
transformers = "~4.16.0"
weaviate-client = "~3.3.0"

[tool.poetry.dev-dependencies]
black = "*"
flake8-aaa = "*"
flake8-absolute-import = "*"
flake8-annotations = "*"
flake8-annotations-complexity = "*"
flake8-bandit = "*"
flake8-breakpoint = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-commas = "*"
flake8-comprehensions = "*"
flake8-docstrings = "*"
flake8-eradicate = "*"
flake8-expression-complexity = "*"
flake8-isort = "*"
flake8-pie = "*"
flake8-print = "*"
flake8-pytest-style = "*"
flake8-return = "*"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.black]
line_length = 120

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
